requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.3.1
alembic==1.7.3
annotated-types==0.6.0
async-timeout==3.0.1
attrs==23.2.0
base58==2.1.0
beautifulsoup4==4.10.0
bitarray==2.9.2
certifi==2021.5.30
cfgv==3.4.0
chardet==4.0.0
charset-normalizer==2.0.6
ckzg==1.0.1
coverage==5.5
curlify==2.2.1
cytoolz==0.12.3
distlib==0.3.8
ens==0.6.1
et-xmlfile==1.1.0
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-tester==0.11.0b2
eth-typing==4.2.2
eth-utils==4.1.0
eth_abi==5.1.0
exceptiongroup==1.2.1
filelock==3.13.4
frozenlist==1.4.1
greenlet==1.1.1
hexbytes==0.2.2
hupper==1.10.3
identify==2.5.36
idna==3.2
importlib-resources==5.2.2
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
Jinja2==3.1.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lru-dict==1.1.7
Mako==1.2.2
MarkupSafe==2.0.1
multiaddr==0.0.9
multidict==5.1.0
netaddr==0.8.0
nodeenv==1.8.0
numpy==1.26.4
openpyxl==3.1.2
packaging==21.0
pandas==2.2.2
parsimonious==0.10.0
PasteDeploy==2.1.1
plaster==1.0
plaster-pastedeploy==0.7
platformdirs==4.2.1
pluggy==1.0.0
pre-commit==3.7.0
protobuf==5.26.1
psycopg2==2.9.1
py==1.10.0
pycryptodome==3.19.1
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.15.1
pyparsing==2.4.7
pyramid==1.10.8
pyramid-debugtoolbar==4.9
pyramid-di==0.4.1
pyramid-jinja2==2.8
pyramid-mako==1.1.0
pyramid-retry==2.1.1
pyramid-services==1.1
pyramid-tm==2.4
pyrsistent==0.18.0
pysha3==1.0.2
pytest==6.2.5
pytest-cov==2.12.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pyunormalize==15.1.0
PyYAML==6.0.1
referencing==0.34.0
regex==2024.4.16
repoze.lru==0.7
requests==2.31.0
rlp==4.0.1
rpds-py==0.18.0
semantic-version==2.10.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==2.0.30
toml==0.10.2
tomli==2.0.1
toolz==0.11.1
transaction==3.0.1
translationstring==1.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.0.7
varint==1.0.2
venusian==3.0.0
virtualenv==20.26.0
waitress==2.1.2
web3==6.18.0
WebOb==1.8.7
websockets==12.0
WebTest==3.0.0
yarl==1.6.3
zipp==3.5.0
zope.deprecation==4.4.0
zope.interface==5.4.0
zope.sqlalchemy==1.6
zstandard==0.22.0
